tests/wpt/tests/navigation-api/navigation-activation/activation-push.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));

  i.contentWindow.navigation.navigate("/common/blank.html?a");
  await new Promise(resolve => i.onload = () => t.step_timeout(resolve, 0));
  let current_entry_after_nav = i.contentWindow.navigation.currentEntry;

  assert_equals(i.contentWindow.navigation.entries().length, 2);
  assert_equals(i.contentWindow.navigation.activation.entry, current_entry_after_nav);
  assert_equals(i.contentWindow.navigation.activation.from,
                i.contentWindow.navigation.entries()[0]);
  assert_equals(i.contentWindow.navigation.activation.navigationType, "push");

  // Same-document navigation doesn't change navigation.activation.
  await i.contentWindow.navigation.navigate("/common/blank.html?a#fragment").finished;

  assert_equals(i.contentWindow.navigation.entries().length, 3);
  assert_equals(i.contentWindow.navigation.activation.entry, current_entry_after_nav);
  assert_equals(i.contentWindow.navigation.activation.from,
                i.contentWindow.navigation.entries()[0]);
  assert_equals(i.contentWindow.navigation.activation.navigationType, "push");
}, "navigation.activation after push");
</script>
